split out ButtonExploreParams
